fixed test square of 42.42

+++ app.rb
@@ -10,7 +10,7 @@ get("/square/new") do
 end
 
 get("/square/results") do
-  @users_number = params.fetch("users_number").to_i
+  @users_number = params.fetch("users_number").to_f
 
   @square = @users_number ** 2.0
